backend/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
requests>=2.31.0
openai>=1.0.0
yt-dlp>=2024.0.0
openai-whisper>=20231117
